requirements.txt
glfw==2.0.0
grpcio==1.38.1
scipy

gt==0.3.14
gtimer==1.0.0b5
gym==0.22.0
h5py==3.1.0
idna==2.10
imageio==2.9.0
joblib==0.17.0
lxml==4.6.2
matplotlib==3.3.2
mujoco-py<2.2,>=2.1
pycparser==2.20
pyglet==1.5.0
PyOpenGL==3.1.5
pyparsing==2.4.7
scikit-learn==0.22
scikit-video==1.1.11
seaborn==0.11.0
serializable==0.2.1
six==1.15.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
threadpoolctl==2.1.0
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tqdm==4.51.0
urllib3==1.26.2
pyyaml==5.4.1
numpy==1.23.0
pandas
path==15.0.0
Pillow
dm_control
envpool
protobuf==3.20.*
PyOpenGL-accelerate
pyrsistent
imageio==2.9.0
imageio-ffmpeg==0.4.4
hydra-core==1.1.0
hydra-submitit-launcher==1.1.5
opencv-python==4.5.3.56
termcolor==1.1.0
POT==0.7.0
rliable